Fix for Array2D windows

+++ Array2D/Data/IReadOnlyArray2D.cs
@@ -6,7 +6,7 @@ namespace BBExtensions.Array2D
     public interface IReadOnlyArray2D<out T> : IEnumerable<T>
     {
         T this[int x, int y] { get; }
-        float ColumnWidth { get; }
+        int ColumnWidth { get; }
         Vector2Int GridSize { get; }
         T[,] ToArray2D();
     }
